Create 09_had.py
change of 'nakresli mapu' to a game 'snake'
snake is 3 signs long at the beginning.
there is a 'food' generated at random position. if 'snake' eats the 'food', it grows - becomes longer. A new 'food' is generated once in 30 game moves. When the snake eats a 'food' and there is no other in the field, a new one is generated.